refactor: move globals and system_settings to utils
