--- src/globals.c
@@ -1,6 +1,0 @@
-#include "../inc/types.h"
-#include "../inc/main.h"
-
-size_t min_chunk_size = 0;
-size_t pointer_size = 0;
-t_heap* global_heap = NULL;
